backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
requests==2.31.0
redis==5.0.4
python-dotenv==1.0.1
Pillow==10.3.0
transformers==4.40.0
torch==2.3.0
pyspellchecker>=0.7.0
celery==5.3.6
fastapi-mail==1.4.1
groq==0.5.0